+++ concrete_datastore/api/v1/views.py
@@ -49,6 +49,7 @@ from concrete_datastore.concrete.models import (  # pylint:disable=E0611
     PasswordChangeToken,
     Email,
     SecureConnectToken,
+    SecureConnectCode,
 )
 from concrete_datastore.api.v1.throttling import (
     CustomUserRateThrottle,
@@ -71,6 +72,7 @@ from concrete_datastore.api.v1.serializers import (
     ConcretePasswordValidator,
     make_serializer_class,
     SecureLoginSerializer,
+    SecureLoginCodeSerializer,
 )
 from concrete_datastore.api.v1.filters import (
     FilterSupportingOrBackend,
@@ -90,7 +92,7 @@ from concrete_datastore.api.v1.filters import (
 
 from concrete_datastore.api.v1.authentication import (
     TokenExpiryAuthentication,
-    expire_secure_token,
+    expire_secure_connect_instance,
     URLTokenExpiryAuthentication,
 )
 from concrete_datastore.concrete.automation.signals import user_logged_in
@@ -233,6 +235,53 @@ class SecurityRulesMixin(object):
         return Response(data=default_options, status=HTTP_200_OK)
 
 
+class RetrieveSecureConnectCode(generics.GenericAPIView):
+    serializer_class = ResetPasswordSerializer
+
+    def post(self, request, *args, **kwargs):
+        # The ResetPasswordSerializer only need an email like this view
+        serializer = ResetPasswordSerializer(data=request.data)
+        if not serializer.is_valid():
+            return Response(
+                data={
+                    "message": serializer.errors,
+                    "_errors": ["INVALID_DATA"],
+                },
+                status=HTTP_400_BAD_REQUEST,
+            )
+
+        email = serializer.validated_data["email"].lower()
+
+        user_queryset = UserModel.objects.filter(email=email, is_active=True)
+        if not user_queryset.exists():
+            return Response(
+                data={
+                    "message": "Wrong email address",
+                    "_errors": ["WRONG_EMAIL_ADDRESS"],
+                },
+                status=HTTP_400_BAD_REQUEST,
+            )
+
+        user = user_queryset.first()
+        # Check if the secure token is expired and expire in database
+        secure_codes = SecureConnectCode.objects.filter(
+            user=user, expired=False
+        )
+        for secure_code in secure_codes:
+            expire_secure_connect_instance(
+                secure_code, settings.SECURE_CONNECT_CODE_EXPIRY_TIME_SECONDS
+            )
+        secure_codes_count = secure_codes.count()
+        if secure_codes_count >= settings.MAX_SECURE_CONNECT_CODES:
+            return Response(status=HTTP_429_TOO_MANY_REQUESTS)
+        secure_connect_code = SecureConnectCode.objects.create(user=user)
+
+        if secure_connect_code.mail_sent is False:
+            secure_connect_code.send_mail()
+        data = {'message': 'Code created and email sent'}
+        return Response(data=data, status=HTTP_201_CREATED)
+
+
 class RetrieveSecureTokenApiView(generics.GenericAPIView):
     """
     This view is used to create a secure token and send an email to the user
@@ -270,7 +319,9 @@ class RetrieveSecureTokenApiView(generics.GenericAPIView):
             user=user, expired=False
         )
         for secure_token in secure_tokens:
-            expire_secure_token(secure_token)
+            expire_secure_connect_instance(
+                secure_token, settings.SECURE_CONNECT_TOKEN_EXPIRY_TIME_SECONDS
+            )
         secure_tokens_count = secure_tokens.count()
         if secure_tokens_count >= settings.MAX_SECURE_CONNECT_TOKENS:
             return Response(status=HTTP_429_TOO_MANY_REQUESTS)
@@ -331,7 +382,9 @@ class GenerateSecureTokenApiView(generics.GenericAPIView):
             user=user, expired=False
         )
         for secure_token in secure_tokens:
-            expire_secure_token(secure_token)
+            expire_secure_connect_instance(
+                secure_token, settings.SECURE_CONNECT_TOKEN_EXPIRY_TIME_SECONDS
+            )
         secure_tokens_count = secure_tokens.count()
         if secure_tokens_count >= settings.MAX_SECURE_CONNECT_TOKENS:
             return Response(status=HTTP_429_TOO_MANY_REQUESTS)
@@ -383,7 +436,10 @@ class SecureLoginApiView(generics.GenericAPIView):
             )
         user = secure_connect_token.user
 
-        if expire_secure_token(secure_connect_token):
+        if expire_secure_connect_instance(
+            secure_connect_token,
+            settings.SECURE_CONNECT_TOKEN_EXPIRY_TIME_SECONDS,
+        ):
             log_request = base_message + (
                 f"Secure login attempt for user {user.email}, "
                 "but the token has expired"
@@ -403,6 +459,79 @@ class SecureLoginApiView(generics.GenericAPIView):
         UserModel.objects.filter(pk=user.pk).update(last_login=timezone.now())
         log_request = base_message + (
             f"Secure login attempt for user {user.email} is successful"
+        )
+        logger_api_auth.info(log_request)
+        return Response(data=serializer.data, status=HTTP_200_OK)
+
+
+class SecureLoginCodeApiView(generics.GenericAPIView):
+    """
+     this view is used to login the user with Secure Login
+     with email and code
+    """
+
+    serializer_class = SecureLoginCodeSerializer
+    api_namespace = DEFAULT_API_NAMESPACE
+
+    def __init__(self, api_namespace, *args, **kwargs):
+        self.api_namespace = api_namespace
+        super().__init__(*args, **kwargs)
+
+    def post(self, request, *args, **kwargs):
+        ip = get_client_ip(request)
+        now = pendulum.now('utc').format(settings.LOGGING['datefmt'])
+        user = self.request.user
+        base_message = f"[{now}|{ip}|{user}|AUTH] "
+
+        serializer = self.serializer_class(data=request.data)
+        if not serializer.is_valid():
+            return Response(
+                data={
+                    "message": serializer.errors,
+                    "_errors": ["INVALID_DATA"],
+                },
+                status=HTTP_400_BAD_REQUEST,
+            )
+        email = serializer.validated_data['email']
+        code = serializer.validated_data['code']
+        try:
+            secure_connect_code = SecureConnectCode.objects.get(
+                user__email=email, value=code
+            )
+        except ObjectDoesNotExist:
+            log_request = base_message + (
+                f"Secure login attempt failed: invalid code"
+            )
+            logger_api_auth.info(log_request)
+            return Response(
+                data={'message': 'Invalid code', "_errors": ["INVALID_CODE"]},
+                status=HTTP_401_UNAUTHORIZED,
+            )
+        user = secure_connect_code.user
+
+        if expire_secure_connect_instance(
+            secure_connect_code,
+            settings.SECURE_CONNECT_CODE_EXPIRY_TIME_SECONDS,
+        ):
+            log_request = base_message + (
+                f"Secure login with code attempt for user {user.email}, "
+                "but the token has expired"
+            )
+            logger_api_auth.info(log_request)
+            return Response(
+                data={
+                    "message": "Code has expired",
+                    "_errors": ["CODE_HAS_EXPIRED"],
+                },
+                status=HTTP_403_FORBIDDEN,
+            )
+
+        serializer = UserSerializer(
+            instance=user, api_namespace=self.api_namespace
+        )
+        UserModel.objects.filter(pk=user.pk).update(last_login=timezone.now())
+        log_request = base_message + (
+            f"Secure login with code attempt for user {user.email} is successful"
         )
         logger_api_auth.info(log_request)
         return Response(data=serializer.data, status=HTTP_200_OK)